feat: change some models feat: add possibility to get newest Exchange Rates from remote API remove: remove unnecessary de/serializers

+++ Reactive-learning-app/order-app/src/main/java/edu/netcracker/order/app/order/utils/OrderUtils.java
@@ -9,9 +9,9 @@ public class OrderUtils {
     private OrderUtils() {
     }
 
-    public static Order postProcessOrderSum(Order order, Function<Float, Float> postProcessPrice, boolean calculateTotalPrice) {
+    public static Order postProcessOrderSum(Order order, Function<Double, Double> postProcessPrice, boolean calculateTotalPrice) {
         if (calculateTotalPrice) {
-            order.setTotalPrice(0f);
+            order.setTotalPrice(0D);
             order.getProducts().forEach(product -> order.setTotalPrice(product.getFirst().getPrice() * product.getSecond() + order.getTotalPrice()));
         }
         order.setTotalPrice(postProcessPrice.apply(order.getTotalPrice()));
